refactored all components to be more reusable
