+++ client/package.json
@@ -14,6 +14,7 @@
     "@reduxjs/toolkit": "^2.2.6",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-icons": "^5.2.1",
     "react-redux": "^9.1.2",
     "react-router-dom": "^6.25.1",
     "redux": "^5.0.1",
