+++ client/src/components/ContentDisplay/ContentDisplay.css
@@ -8,14 +8,14 @@
 .content-display__sidebar{
     display: flex;
     flex-direction: column;
-    width: 20%;
+    width: 10%;
     height: 100%;
 }
 
 .content-display__main{
     display: flex;
     flex-direction: column;
-    width: 80%;
+    flex-grow: 1;
     height: 100%;
     border: red solid 1px;
 }
