+++ client/src/components/Sidebar/Sidebar.css
@@ -1,5 +1,21 @@
 .sidebar__main {
     display: flex;
     flex-direction: column;
+    align-items: center;
+    color: gray;
+    height: 100%;
+
+
+
     border: 1px solid yellow;
+}
+
+.sidebar__top {
+    display: flex;
+    align-items: center;
+    padding: 20px 15px;
+}
+
+.sidebar__logo {
+    font-size: 25px;
 }
