+++ client/src/components/Sidebar/Sidebar.jsx
@@ -1,13 +1,19 @@
 import { useState } from "react";
 import { NavLink } from "react-router-dom";
-import "./Sidebar.css"
+import "./Sidebar.css";
+import { FaBars } from "react-icons/fa";
 
 export default function Sidebar({ sidebarData }) {
   const [isExpanded, setIsExpanded] = useState();
 
   return (
     <div className="sidebar__main">
-      <div className="sidebar__top">DPS.GG</div>
+      <div className="sidebar__top">
+        <h1 className="sidebar__top-name">DPS.GG</h1>
+        <div className="sidebar__top-logo">
+          <FaBars />
+        </div>
+      </div>
       {sidebarData.map((option, index) => {
         return (
           <NavLink to={option.path} className="sidebar__link" key={index}>
