+++ client/src/data/NavbarOptions.jsx
@@ -1,0 +1,15 @@
+import LeagueOfLegendsIcon from "../assets/LeagueOfLegendsIcon.png"
+import OsuIcon from "../assets/OsuIcon.png"
+
+export const navbarData = [
+    {
+        path: "/lol",
+        name: "League Of Legends",
+        icon: LeagueOfLegendsIcon
+    },
+    {
+        path: "/osu",
+        name: "Osu!",
+        icon: OsuIcon
+    }
+]
