--- client/src/data/SidebarOptions.js
@@ -1,20 +1,0 @@
-export const lolSidebarData = [
-    {
-        path: "/lol",
-        name: "Home",
-        icon: "House Icon"
-    },
-    {
-        path: "/tier-list",
-        name: "Tier List",
-        icon: "three little lines stacked/paper"
-    }
-]
-
-export const osuSidebarData = [
-    {
-        path: "/osu",
-        name: "Home",
-        icon: "House Icon"
-    }
-]
