+++ client/src/data/SidebarOptions.jsx
@@ -1,0 +1,22 @@
+import { FaHome } from "react-icons/fa";
+
+export const lolSidebarData = [
+  {
+    path: "/lol",
+    name: "Home",
+    icon: <FaHome />,
+  },
+  {
+    path: "/tier-list",
+    name: "Tier List",
+    icon: <FaHome />,
+  },
+];
+
+export const osuSidebarData = [
+  {
+    path: "/osu",
+    name: "Home",
+    icon: <FaHome />,
+  },
+];
